chore: rename config.js to constants.js
