+++ src/utils.js
@@ -1,4 +1,4 @@
-import { CONFIG } from './config'
+import { CONFIG } from './constants'
 
 /**
  * Generic utility which merges two objects.
